[JENKINS-26694] Make tests succeed on Windows: HudsonTestCase using workflow-1.1 fails for JENKINS-26030.
